move is_subzone to the root of the recursor crate
