+++ crates/recursor/src/recursor.rs
@@ -14,9 +14,6 @@ use lru_cache::LruCache;
 use parking_lot::Mutex;
 use tracing::{debug, info, warn};
 
-#[cfg(test)]
-use std::str::FromStr;
-
 use crate::{
     proto::{
         op::Query,
@@ -391,7 +388,7 @@ impl Recursor {
                     .chain(r.take_name_servers())
                     .chain(r.take_additionals())
                     .filter(|x| {
-                        if !is_subzone(ns.zone().clone(), x.name().clone()) {
+                        if !super::is_subzone(ns.zone().clone(), x.name().clone()) {
                             warn!(
                                 "Dropping out of bailiwick record {x} for zone {}",
                                 ns.zone().clone()
@@ -457,7 +454,7 @@ impl Recursor {
                 //     .filter_map(Record::data)
                 //     .filter_map(RData::to_ip_addr);
 
-                if !is_subzone(zone.base_name().clone(), zns.name().clone()) {
+                if !super::is_subzone(zone.base_name().clone(), zns.name().clone()) {
                     warn!(
                         "Dropping out of bailiwick record for {:?} with parent {:?}",
                         zns.name().clone(),
@@ -589,90 +586,6 @@ fn recursor_opts() -> ResolverOpts {
     options.num_concurrent_reqs = 1;
 
     options
-}
-
-/// Bailiwick/sub zone checking.
-///
-/// # Overview
-///
-/// This function checks that two host names have a parent/child relationship, but does so more strictly than elsewhere in the libraries
-/// (see implementation notes.)
-///
-/// A resolver should not return answers outside of its delegated authority -- if we receive a delegation from the root servers for
-/// "example.com", that server should only return answers related to example.com or a sub-domain thereof.  Note that record data may point
-/// to out-of-bailwick records (e.g., example.com could return a CNAME record for www.example.com that points to example.cdnprovider.net,)
-/// but it should not return a record name that is out-of-bailiwick (e.g., we ask for www.example.com and it returns www.otherdomain.com.)
-///
-/// Out-of-bailiwick responses have been used in cache poisoning attacks.
-///
-/// ## Examples
-///
-/// | Parent       | Child                | Expected Result                                                  |
-/// |--------------|----------------------|------------------------------------------------------------------|
-/// | .            | com.                 | In-bailiwick (true)                                              |
-/// | com.         | example.net.         | Out-of-bailiwick (false)                                         |
-/// | example.com. | www.example.com.     | In-bailiwick (true)                                              |
-/// | example.com. | www.otherdomain.com. | Out-of-bailiwick (false)                                         |
-/// | example.com  | www.example.com.     | Out-of-bailiwick (false, note the parent is not fully qualified) |
-///
-/// # Implementation Notes
-///
-/// * This function is nominally a wrapper around Name::zone_of, with two additional checks:
-/// * If the caller doesn't provide a parent at all, we'll return false.
-/// * If the domains have mixed qualification -- that is, if one is fully-qualified and the other partially-qualified, we'll return
-///    false.
-///
-/// # References
-///
-/// * [RFC 8499](https://datatracker.ietf.org/doc/html/rfc8499) -- DNS Terminology (see page 25)
-/// * [The Hitchiker's Guide to DNS Cache Poisoning](https://www.cs.utexas.edu/%7Eshmat/shmat_securecomm10.pdf) -- for a more in-depth
-///   discussion of DNS cache poisoning attacks, see section 4, specifically, for a discussion of the Bailiwick rule.
-fn is_subzone(parent: Name, child: Name) -> bool {
-    if parent.is_empty() {
-        return false;
-    }
-
-    if (parent.is_fqdn() && !child.is_fqdn()) || (!parent.is_fqdn() && child.is_fqdn()) {
-        return false;
-    }
-
-    parent.zone_of(&child)
-}
-
-#[test]
-fn is_subzone_test() {
-    assert!(is_subzone(
-        Name::from_str(".").unwrap(),
-        Name::from_str("com.").unwrap()
-    ));
-    assert!(is_subzone(
-        Name::from_str("com.").unwrap(),
-        Name::from_str("example.com.").unwrap()
-    ));
-    assert!(is_subzone(
-        Name::from_str("example.com.").unwrap(),
-        Name::from_str("host.example.com.").unwrap()
-    ));
-    assert!(is_subzone(
-        Name::from_str("example.com.").unwrap(),
-        Name::from_str("host.multilevel.example.com.").unwrap()
-    ));
-    assert!(!is_subzone(
-        Name::from_str("").unwrap(),
-        Name::from_str("example.com.").unwrap()
-    ));
-    assert!(!is_subzone(
-        Name::from_str("com.").unwrap(),
-        Name::from_str("example.net.").unwrap()
-    ));
-    assert!(!is_subzone(
-        Name::from_str("example.com.").unwrap(),
-        Name::from_str("otherdomain.com.").unwrap()
-    ));
-    assert!(!is_subzone(
-        Name::from_str("com").unwrap(),
-        Name::from_str("example.com.").unwrap()
-    ));
 }
 
 #[tokio::test]
